refactor: merge ask.py into personality.py
