+++ cogniq/personalities/chat_gpt4/chat_gpt4.py
@@ -8,10 +8,7 @@ logger = logging.getLogger(__name__)
 
 from cogniq.personalities import BasePersonality
 from cogniq.slack import CogniqSlack
-from cogniq.openai import CogniqOpenAI
-
-
-from .ask import Ask
+from cogniq.openai import system_message, user_message, CogniqOpenAI
 
 
 class ChatGPT4(BasePersonality):
@@ -32,13 +29,11 @@ class ChatGPT4(BasePersonality):
         self.cslack = cslack
         self.copenai = copenai
 
-        self.ask = Ask(cslack=cslack, copenai=copenai)
-
     async def async_setup(self) -> None:
         """
         Please call after initializing the personality.
         """
-        await self.ask.async_setup()
+        pass
 
     async def ask_task(self, *, event: Dict, reply_ts: float, context: Dict) -> None:
         channel = event["channel"]
@@ -50,7 +45,7 @@ class ChatGPT4(BasePersonality):
         history = await self.cslack.openai_history.get_history(event=event, context=context)
         # logger.debug(f"history: {history}")
 
-        ask_response = await self.ask.ask(q=message, message_history=history, context=context)
+        ask_response = await self.ask(q=message, message_history=history, context=context)
         await self.cslack.chat_update(channel=channel, ts=reply_ts, context=context, text=ask_response["answer"])
 
     async def ask_directly(
@@ -65,8 +60,45 @@ class ChatGPT4(BasePersonality):
         """
         Ask directly to the personality.
         """
-        ask_response = await self.ask.ask(q=q, message_history=message_history, stream_callback=stream_callback, context=context)
+        ask_response = await self.ask(q=q, message_history=message_history, stream_callback=stream_callback, context=context)
         return ask_response["answer"]
+
+    async def ask(
+        self,
+        *,
+        q: str,
+        message_history: List[Dict[str, str]],
+        stream_callback: Callable[..., None] | None = None,
+        context: Dict,
+        reply_ts: float | None = None,
+    ) -> Dict[str, Any]:
+        if message_history is None:
+            message_history = []
+        # bot_id = await self.cslack.openai_history.get_bot_user_id(context=context)
+        bot_name = await self.cslack.openai_history.get_bot_name(context=context)
+        # logger.info(f"Answering: {q}")
+
+        # if the history is too long, summarize it
+        message_history = self.copenai.summarizer.ceil_history(message_history)
+
+        # Set the system message
+        message_history = [system_message(f"Hello, I am {bot_name}. I am a slack bot that can answer your questions.")] + message_history
+
+        # if prompt is too long, summarize it
+        short_q = await self.copenai.summarizer.ceil_prompt(q)
+
+        logger.info("short_q: " + short_q)
+        message_history.append(user_message(short_q))
+
+        res = await self.copenai.async_chat_completion_create(
+            messages=message_history,
+            stream_callback=stream_callback,
+            model="gpt-4",  # [gpt-4-32k, gpt-4, gpt-3.5-turbo]
+        )
+
+        answer = res["choices"][0]["message"]["content"]
+        logger.info(f"final_answer: {answer}")
+        return {"answer": answer, "response": res}
 
     @property
     def description(self) -> str:
